web3 connection setup added
