--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-	preset: "ts-jest",
-	testEnvironment: "node",
-	modulePathIgnorePatterns: ["<rootDir>/src/__tests__/utils"]
-};
