+++ tsconfig.json
@@ -1,6 +1,5 @@
 {
 	"compilerOptions": {
-		// "lib": ["ES2020"],
 		"module": "CommonJS",
 		"target": "ES2017",
 		"outDir": "dist/",
@@ -15,5 +14,5 @@
 			"typings": ["typings"]
 		}
 	},
-	"exclude": ["dist", "src/__tests__", "jest.config.js"]
+	"exclude": ["dist", "src/test", "jest.config.js"]
 }
